add disposable component base
