+++ src/Ozds.Client/Components/Providers/CultureStateProvider.razor
@@ -1,6 +1,7 @@
 @namespace Ozds.Client.Components.Providers
+@using Ozds.Client.Components.Base
 
-@inherits Ozds.Client.Components.Base.OzdsComponentBase
+@inherits DisposableComponentBase
 
 @if (_state is null)
 {
